+++ tests/load_tests/opensearch_related_article_test.py
@@ -3,7 +3,7 @@ import random
 from pathlib import Path
 from locust import HttpUser, task
 
-from sciety_labs.providers.opensearch_module.opensearch import OpenSearchConnectionConfig
+from sciety_labs.providers.opensearch.opensearch import OpenSearchConnectionConfig
 from tests.load_tests.data import get_data_file_path
